ldapadmin - do not validate client-side - it must be done server-side
The tests are done client-side at the same time the user is filling the
form. With the previous client-sde validation, the Accept button appeared
to de disabled, because clicking had no effect, even no erreor message.

Important: the validation must be done server-side. This is NOT the case
currently for the password strength check.