Add `publicContextPath` attribute to JSP to forge absolute CSS and JS URIs
In JSP pages, the CSS files, for example, are changed from

```
	<link href='../account/css/bootstrap.min.css' rel="stylesheet" />
```

to

```
	<link href='${publicContextPath}/account/css/bootstrap.min.css' rel="stylesheet" />
```

ie. to

```
	<link href='/ldapadmin/account/css/bootstrap.min.css' rel="stylesheet" />
```

This way, the JSP files may be called for any URI depth level.

Note that we used an interceptor to add an attribute coming from the geOrchestra configuration,
the `ldapadmin/` publicContextPath. We could not use the common `${pageContext.request.requestURL}`
attribute in JSP files since it's equal to `ldapadmin-private` (name of the webapp) in
the default geOrchestra installation.