Mapfishapp - Bumping to newer version of spring
There is a bug in previous versions of spring, which is not able to
resolve a setter while instanciating a bean. My previous modifications
were using the version of Spring defined in the root pom, this fixes for
mapfishapp only a newer version (which is currently used in production
on the Rennes-métropole platform).