ldapadmin - Revamp on configuration tweaking management
Doing substitutions into a property file which is sourced by spring
directly into the XML file (as it is actually done on the security-proxy
/ downloadform ...).

Begining to rewrite UserTokenDao so that it uses a connection pool,
instead of opening a PG connection each time it is needed (bug prone).
Ensuring that the connection is closed after use.

TODO: this has been done before the will to add some utests, completing
the testsuite. So, next work is to test it.