ldapadmin - fixing pigma issue #131
In order to compare equality of 2 users from the LDAP, the sn field were
used, which sounds a bad idea, because it is usually used to fill in the
last name of the users, and causing a bug (users not retrieved) when 2
member of the organization were sharing the same last names. We shall
use Account uid (uid field ldap-side) instead.

This commit introduces a new dependency for testing related to
collecting errors, to avoid premature end of the test, so that we can
use the UsersGroupsControllerTest class to get the inconcistencies of a
LDAP tree, using e.g.:

$ mvn clean test -Dtest=UsersGroupsControllerTest \
-Dldapadmin.test.openldap.activated=true \
-Dldapadmin.test.openldap.binddn="cn=admin,dc=georchestra,dc=org" \
-Dldapadmin.test.openldap.password=secret \
-Dldapadmin.test.openldap.ldapurl="ldap://localhost:389" \
-Dldapadmin.test.openldap.basedn="dc=georchestra,dc=org"

In case of errors, the output can be as follows:

  testUsersGroupController(org.georchestra.ldapadmin.ws.backoffice.UsersGroupsControllerTest):
  cs40_mimizan is not in the expected users, but is member of groups(..)
  testUsersGroupController(org.georchestra.ldapadmin.ws.backoffice.UsersGroupsControllerTest):
  castel is not in the expected users, but is member of groups(..)
[...]
  testUsersGroupController(org.georchestra.ldapadmin.ws.backoffice.UsersGroupsControllerTest):
  pigmaopendata is not in any groups(..)
  testUsersGroupController(org.georchestra.ldapadmin.ws.backoffice.UsersGroupsControllerTest):
  lducos  is not in any groups(..)
[...]

This commit also fixes another bug related of the fakeuser whitelisting,
which is kind of linked to the filed issue.