ldapadmin - fixes #650
This fix is inspired by the current work on the testsuite, but some of
the work could not be fully backported (for conflicting reasons /
introduced features in master and not in 14.01). This has only been
tested at runtime, to ensure that the bug is fixed, but if possible
consider migrate to a more recent version of LDAPAdmin (master is much
more thoroughly tested and probably safer than this one).