mapfishapp - dynamic WMC controller
This is kind of related to issue #40 : We set a large json array in
GEOR_Custom.js while this could be calculated dynamically by a
webservice.

This commit is a backport of a current work which aims to export
configuration of some aspects outside of the webapp. Since this can be
easily extracted and proposed onto master earlier than planned, let's go
for it.

Tests: unit tests backported, runtime tests run.

TODO: we might want to return the default array currently provided by
the configuration instead of an empty JSON array.

Note: since this is not currently wired up to the interface, merging it
into master is pretty harmless.