extractorapp - datadir
This one was one of the first webapp when I began to "datadirize"
geOrchestra. Hence, some better approach with less Java code
modifications could be used (by sourcing the properties file as it is
done in the other webapps, in the ws-servlet.xml).

This is a first shot, but would deserve a little rework.