ldapadmin - administrators can modify users' uid
Evolution funded by CIGALsace
(https://github.com/camptocamp/georchestra-cigalsace-configuration/issues/531)

Enables the ability for LDAP administrators (members of MOD_LDAPADMIN)
to modify the user id (uid) of the users LDAP entries.

The uid field is not like the others attributes and need extra work to
implement the ability to modify it, since it is part of the DN
qualification ; it requires to "move" the LDAP entry, which is not the
same as updating a regular attribute.

It also requires to get the users groups and update the member entry.

Tests: Utests + tests against a real ldap adapted / OK
Runtime tests OK