server-deploy - removing classifiers
since the 15.12 version, the naming convention of the generated wars
changed a bit. The classifier is not used anymore, making server-deploy
unable to find the artifacts back into the local .m2 repository.

Caveats: This is no longer possible to have multiple wars for different
configurations in the same m2 repository, since the last compiled webapp
will be used.