SP - redirect every webapps to webapps/
This was previously done by the frontend webserver configuration. But in
some configurations (docker) there is no need to introduce an extra
frontend webserver. This assures that the webapps have the trailing / on
the path, sending a redirect to the client, e.g.:

- http://security-proxy/geonetwork -> /geonetwork/
- /geoserver -> /geoserver/
- ...

Following the comments on the PR, there is no need to adapt the
documentation because it is still required that Apache (if there is a
one) adds it if it intercepts the query before sending it to the SP.

Tests: runtime.