added the sudo
oh i forgot the sudo command... sorry !
but for the point, my opinion is that it is a good thing: you want to copy files after the /, at the point position: "dir/."
so, it is a good practice, easy to understand, without ambiguities. 
it will go over my body if you want me to remove the point. ;)