root pom - adding a profile for jsonevent dep
This adds an optional dependency in the webapps to be able to log in a
JSON fashion compatible with logstash and ELK stacks.

See https://github.com/logstash/log4j-jsonevent-layout#usage for more
information about this feature.

This should not affect regular builds and requires the extra profile to
be activated so that the dependency is added to the generated wars (for
now on, only geonetwork should not be affected by this commit).

Tests: runtime tested on mapfishapp. since the added dependencies are
not related to currently used libraries, this should not harm the
current behaviour.