postgresql docker - fixing patch for the entrypoint
The build on jenkins is failing since a probable change upstream on the
postgresql entrypoint. The patch needs to be adjusted.

Anyway, even if I am pretty confident in this patch, I'd prefer to go
through a review since if something changed upstream it might be a
reason.

Tests: docker build OK, but might deserve a runtime test.