extractorapp - "fixing" testsuite when run as root
See comment in the original test ; this test has not been written with
the possibility it could run as root one day. Well, in a docker
environment, this can actually happen. This fix is far from perfect, see
the following comment:

https://github.com/georchestra/ci-jenkinsfiles/issues/1#issuecomment-275871254

but should at least avoid a false positive in the testsuite execution.

Tests:
- non-privileged user OK
- as root             OK