Forgot password, try to avoid stack traces. Stack traces are erratic.
Don't know how to test it.

Suppose it has to do with attempt to reuse yesterday connection (it happens in the morning ?), wich are not discarded from the connection pool.

Apache docs states that when "testOnBorrow" true, validation query is "The SQL query that will be used to validate connections from this pool before returning them to the caller. If specified, this query MUST be an SQL SELECT statement that returns at least one row. If not specified, connections will be validation by calling the isValid() method".

Theory is that isValid method is not sufficient to leverage the connection up.