Add index on ogc stats tables
This commit modify trigger in order to automaticly add indexes on stats
tables. It will create index when 'rotating' logs: when a new table is
created, indexes will be added to previous table.
This commit also remove patch of postgres entrypoint. Instead, a script
was added to clear *-data.sql file if IGNORE_DATA=yes.