SP - fixing wicket 7.6.0 / GS 2.12.1 behaviour
The SP when forwarding a POST request will in certain cases strip the
extra parameters passed to the target url, e.g.:

POST /geoserver/web/wicket?blah will actually be posted onto
/geoserver/web/wicket

This results in having a GeoServer UI broken, since some wicket logic
are missing. this might also have broken other things, but skimming the
git history, I could not find any valid reason to strip the URL as it is
done, so I removed the code and runtime tested, it does not seem to
break anything and will also simplify the SP logic.