GN 3.4.1 / GS 2.12 integration
= GS integration =

This commit partially reverts 4400757aa87ed3baf9423c84af977e806b2bf4c4
which introduced a regression. There are 2 types of POST request:

* the request body is in a URLencoded form,
* the request body is a payload

The second type is mainly used in some applications (LDAPAdmin), where
the body is a JSON payload. We need to send it, the previous commit was
stripping the body, hence the POST data was null.

= GN integration =

In order to protect the webapp for Cross-site-request-forgery (CSRF)
attacks, a token is computed which can be fetch from a POST call onto
/geonetwork/srv/{lang}/xml.info. The ajax call leads to a 403 because of
the missing CSRF token, but the header Set-Cookie provides the token,
but the SP was not transmitting this header to the browser. This commit
ensures that a webapp providing such a header is correctly provided to
the actual client.

Tests: Tested on the GGE instance.