GS2.12 - geofence - exclude gf-model
2 jars are providing the same classes:
* gf-model
* gf-model-internal

The first one is only used in standalone mode for geofence, which is not
the mode used in geOrchestra (see
https://github.com/geoserver/geofence#architecture).

Both provide the same classes, so we have to exclude the first one from
the georchestra build.

Tests: tested on GGE, we are now able to add rules. Unfortunately, h2db
seems to be used for now. Still work to be done on the GF configuration
so that PostGreSQL is used.