Makefile - fixing GS build
Sorry, I forgot about this makefile. the pom.xml in geoserver/ now
references the modules needed by the build, depending on which profile
is activated (mainly geofence or not geofence).

So depending on the chosen build mode, it will generate the specific
georchestra geoserver jars in the geoserver-submodule/ subdirectories,
which should concern 3 or 4 jars max, the rest comes from the official
boundless repositories.

To avoid ambiguity, I only kept the geofence profile in poms, no
geofence-server profile anymore, but the needed dependencies should
finish in the generated artifact in the end.