Console - Fix loading of email configuration
It seems that we cannot populate a Properties and then construct Session
object with the Properties instance. So we need to first create Session
instance and then use setProperty() method on it.

It also seems that we need to use a property called 'mail.smtp.port'
instead of 'mail.protocol.port'.

Fix #2011