Fix a bug in console role management by which isFavorite:false won't be updated
Setting a role as favorite works, but setting it back to false
doesn't. This patch fixes the issue. Note though that proper
testing for the fix requires some integration testing infrastructure
which is being worked out on the 18.12 branch as it would
be overkill to include on this one this close to a release.