ogc-service-statistics: allow to set a DataSource externally
DataServicesConfiguration is a singleton that was maintaining
a single sql Connection as an instance variable to be reused
by all `DataCommand`s, which produced errors when the connection
got closed.

Now it'll create a PGSimpleDataSource and clients of
DataServicesConfiguration.getConnection() are required to close
the returned connection, following JDBC best practices.

A pooing DataSource can be given, which is what proxy will do,
allowing it to follow the homogenized connection pool configuration
that applies to all geOrchestra applications.