security-proxy: configure the OGC stats logger DataSource
Adds a standard way of configuring the connection pool
used for the OGC stats logger appender from the ogc-server-statistics
module.

The DataSource is configured with the `ogcStatsDataSource` name
instead of the default `dataSource` name because the application
might want to use the default data source name for its own business
in the future.