Add fixed geotools s3-geotiff jar file to GeoServer Docker image
Current version of gt-s3-geotiff tried to auto configure
its ehcache programmatically and conflicted with any other
module that uses ehcache (for instance, geofence) as it tried
to create a defaultCache that might already exist.

Also, the option of passing an ehcache xml configuration file
location as a System property didn't work, and was fixed in
GeoTools master (22-SNAPSHOT) as of PR 2308, and backported to
21.x, 20.x, 19.x. Since we're on GeoTools 18.5 and the 18.x
series is end of life, there won't be another release from it,
this patch adds the fixed jar file directly when creating the
georchestra geoserver docker image.

The fixed jar file is located under
geoserver/lib/2.12.5/gt-s3-geotiff-18.5.jar, as an exceptional
way of replacing a geoserver dependency (otherwise we'd need to
add geotools 18.x as a submodule, build it ourselves, and still
replace the 18.5 library by the 18-SNAPSHOT one).

See the comments on geoserver/lib/README.md for the instructions
on how to add these kind of exceptions to override a geoserver
dependency based on the specifig geoserver version being used.