ldap - create a LDAP schema for geOrchestra
Create a new LDAP schema to be imported in the OpenLDAP server.

The schema is named cn=georchestra,cn=scheam,cn=config, and it's aimed
at defining the objectClasses and attributes required by geOrchestra.

The first elements in that schema are:

- georchestraUser objectClass: it's an AUXILIARY objectClass, that means
it can be added to an existing object (typically an inetOrgPerson inside
ou=users,dc=georchestra,dc=org), to add attributes
- privacyPolicyAgreementTime attribute: it's an attribute of type
GeneralizedTime, which purpose is to store the date and time when a
geOrchestra user has agreed to the terms and conditions of the platform.
In particular, it's meant to provide a mean to comply with the European
General Data Protection Regulation (RGPD).

Important: the OID of both definitions are TEMPORARY and MUST be
changed, once the geOrchestra project will get an OID assigned.