console - add a property to Account for conditions agreement
Due to RGPD, there is a need to keep track of if and when a geOrchestra
user has agreed to the privacy policy (or more generally to the terms
and conditions).
We use LocalDate type to encode that date. One day is a sufficient
granularity. Note that it implies the use of Java 8, at least.
The date is serialized/deserialized with string format "2019-02-25".