GS - fix versions for JPA / Hibernate / spring-orm with geofence
geofence-persistence requires a spring version (esp. spring-orm)  which
is incompatible with the spring version needed by GeoServer. We force
for this very specific module (with the profile activated) the
dependencies in the correct versions so that GS can boot.

Note: We might encounter some issues mixing the dependency versions this
way, but since JPA mechanisms in GeoServer seem only used in the very
specific case of geofence, we can at least try it out.