pom - Adding jacoco as code coverage tool optionnally
Some external code analysis tools used in the community (deutsch telekom
do have a SonarQube deployed in their infrastructure, we have a
communautary Jenkins) can make use of the generated report by Jacoco
to get insights on the code coverage of the solution.

This commit proposes to activate the maven plugin as an optional profile
which can (or not) be activated.

Tested in the security-proxy as:

```
$ mvn clean test -PjacocoCoverage
```

This generates a target/site/jacoco path.

Without:

```
$ mvn clean test
```

No code coverage report is generated, the behaviour is exactly the same
as currently in master.