mapfishapp: make GeoJSON conversion streaming and avoid OOM
Calls to /togeojson were building the wholse GeoJSON
content on a StringBuilder and then to a String.

This patch makes the process streaming, saving the whole
response to a temporary file without loading the full
feature collection in memory nor building the response
in memory, and then streaming to the http response.

Also improves exception handling