build: configure formatter-maven-plugin
Add the net.revelc.code.formatter:formatter-maven-plugin
maven plugin to the root pom's pluginManagement
section and configure it with to be inherited
by child projects that add it to their <plugins> section.

The plugin is configured to format only java source
code files using the .mvn/formatter.xml eclipse
formatter definition.

This formatter can be directly imported from eclipse IDE
and from IntelliJ IDEA's "Eclipse Code Formatter" plugin
(https://plugins.jetbrains.com/plugin/6546-eclipse-code-formatter).

When compiling a project that includes the formatter-maven-plugin,
the default action is to format the source code to adhere to the
defined code style.

The other option is to instruct the plugin to check for correct
formatting and fail the build.

This behavior is controlled by the fmt.action property, like in:
mvn compile -Dfmt.action=format or -Dfmt.action=validate.

On a CI build server (Jenkins/Travis-CI), the -Dfmt.action=validate
should be used, to ensure a committer did his/her due diligence
before pushing to the upstream git repository.

Finally, the plugin can be disabled with the -Dfmt.skip=true
maven property.