Console - Fixing GDPR download w/ geonetwork
Dates in GeoNetwork's table metadata are stored as text / varchar
fields, and follow the iso8601 spec (mainly a 'T' instead of a space to
separate the date from the hours).

The current logic in the code was to try to parse the invert (space
instead of 'T') for dates coming from GeoNetwork.

Tests: added a utest to specifically parse using the GEONETWORK_DATE_FORMAT constant.
utests ok (and adapted consequently).
Runtime tested directly by injecting the class onto dev.geo2france.fr