console - do not change the logo if empty
The current code always tries to change the logo in the LDAP, even if
Spring receives an empty file. So when a user change a field but without
sending a logo, it discards the jpegPhoto attribute of the organisation.

Skipping the setLogo() logic when the received file is empty.

Tests:
* no test added
* utest ok
* IT ok
* runtime tested via JDWP and java code switching directly from
  Eclipse.