console - (re ?)introducing urlrewrite (#2449)
See issue #2449 for the motivation.

Caveats: using path-info to get the request-uri without the context name
did not seem to work at runtime, so I used the request URI but with a
wildcard pattern on the webapp name (see in urlrewrite.xml). As a
consequence, "/console/manager/public/libraries.css" will match as well
as "/aaaa/bbbb/manager/public/libraries.css". But at least, it shall
allow people deploying in an other context name to still trigger the
rule accordingly.

tests:

* utests OK
* IT OK
* runtime tests OK