console - using the http headers instead of LDAP (#2983)
Instead of using an extra LDAP connection to retrieve the orgs and roles
of the current logged user, guess the infos from the HTTP headers
received from the security-proxy.

See https://github.com/georchestra/georchestra/issues/2983.

Tests:

* utests ok
* IT ok

Note:

Since the connection to an OpenLDAP is not necessary anymore, we don't
rely on it in the tests. That is the reason why I removed the IT which
was testing the profile endpoint, and migrated to some new unit tests
instead.