gh actions - activating sentry-spring & log4j-logstash profiles
Note:

This will add optional dependencies to the generated docker images. The
overhead is ~ 750 KB, so even if these dependencies are not used at
runtime, we consider that the size increase is acceptable, compared to
the artifact actual size.

For the record:

* log4j-logstash adds the jsonevent-layout log4j1 dependency, which
allows to have a "jsonified" output, which is interesting in case of
parsing the logs with tools like logstash.

* sentry-spring will add the following ones:

```
[INFO] \- io.sentry:sentry-spring:jar:1.7.27:compile
[INFO]    \- io.sentry:sentry:jar:1.7.27:compile
```

And allows the webapp to be connected against a sentry instance. See
https://docs.sentry.io/clients/java/integrations/#spring for
the configuration details.