all webapps - remove sentry-spring / reintroduce sentry-log4j
Quoting the sentry doc:

A Sentry logging integration is more general and will capture errors
(and possibly warnings, depending on your configuration) that occur
inside or outside of a Spring controller. In most scenarios, using one
of the logging integrations instead of sentry-spring is preferred.

https://docs.sentry.io/clients/java/integrations/#spring

Which is actually what we did in the first place.