console - re-enable passwordRecovery endpoint
Something is unclear to me: There are 2 options to arrive to this page:

* from the CAS, being not connected, clicking onto the "lost password?"
  link
* from the console, being connected as admin, then willing to reset the
  password of a user.

In the first case, the email parameter passed via GET is blank. In the
second one, it is set to the email from the user concerned by the
password reset procedure.

for now, only the administrator can access via the second option, as
there is no wildcard here:
https://github.com/georchestra/datadir/blob/master/security-proxy/security-mappings.xml#L17

In this commit, I will keep it this way, as it is the closest from the
current behaviour.

Tests: runtime