Allow to change geoserver_privileged_user's password on first run
If we change his password in the georchestra_datadir, we also have to change it
in the LDAP. This allows to pass it as a secret. The mounted path is given
in the GS_MASTER_PASSWORD_FILE env. var.
It only works on first runs. If the initialization has already taken place
it won't change the password.