Disable loading properties from georchestra datadir in the 'local' profile
This might be a temporary measure to get things up and running.
If not disabling the datadir in the local profile, the geoserver backend
datastore configuration properties, being a simple Map<String, String>,
get mixed up from both locations, datafeeder.properties and
application.yml, leading to an error in geoserver (in this case not
finding the jndi name configured in datafeeder.properties, when the
intention is to completely override the parameters in application.yml's
'local' profile).