Fix race condition in usage of FilterBasedLdapUserSearch
`LdapUserDetailsRequestHeaderProvider` used the same instance of
`FilterBasedLdapUserSearch` with and without setting
`setReturningAttributes`. With both being singletons, there's a race
condition. `FilterBasedLdapUserSearch` is a lightweight object and
afe to be instantiated everytime it's needed though, doesn't need
to be a singleton, the heavy object is the `BaseLdapPathContextSource`
it gets as constructor argument.