datafeeder-ui - serves the index.html file on dynamic endpoints
the JS app will manage the actual requested path in some cases, e.g.  if
/import/<uuid>/validation is requested, then we have to return the
index.html content to the client, which will parse JS-side the url
fragment and display the according page.

The behaviour should be slightly the same as the suggested nginx
configuration from the geonetwork-ui repository, see:

https://github.com/georchestra/geonetwork-ui/blob/redirect-to-index-html/nginx-default.conf

Note: I first tried to do it in a more "bare-metal servlet" approach,
thinking that the servlet-mapping's url-pattern parameters in the
web.xml would suffice, but it actually does not support regexps and is
more of a simple globbing. Then the "fastest" way to figure out how I
could send whatever I wanted to the client was to do spring-mvc, and
introducing spring as a dependency. Maybe a urlrewrite would have been
lighter.

Note2: I could have converted the project into a spring-boot one, but
the idea is still to finish with a webapp war that could be easily
deployed into an existing servlet container for non-dockerized approach.

Tests: mainly runtime-tested via `mvn clean package jetty:run` with
loads of curl.