[datafeeder] Publish only the requested dataset when several where uploaded
If the user uploaded a zipfile with several datasets (e.g. shapefiles),
then the publish request is executed, it contains the list of datasets
to publish among their metadata.

Currently the datafeeder-ui expects an uploaded zip file to contain a
single shapefile, but nothing prevents the user from providing a zip
file with several shapefiles. This backend application is prepared to
handle multiple datasets publications in a single call though.