[datafeeder] Fix single line multiple variables resolution on email templates
Fix variable names resolution when there is more than one variable on
the same line of email templates.

Change templates to be closer to what the real ones shall be.

Better handle job failures.  We need to leave the spring-batch job complete normally or the
publishStatus/error won't be persisted, as spring-batch and our JPA
repositories share the same entity manager. If a publishing process failed and
the exception was re-thrown, spring-batch would abort the current
transaction.