[datafeeder] Fix mismatch between PostGIS table name and requested GeoServer native layer name
Datasets with mixed case were failing to be published to GeoServer with
an error "Trying to create new feature type inside the store, but no
attributes were specified".

Thouhg a bit cryptic, it means that it couldn't find the requested table
name in PostGIS when processing the POST request for a FeatureTypeInfo,
from its `nativeName` property, and was trying to create such table, but
the FeatureTypeInfo provided no AttributeTypeInfo's.

The root of the issue was that Datafeeder was using the "dataset native
name" instead of the "publishing imported name" to set the
FeatureTypeInfo.nativeName property (i.e., it was passing the type name
of the uploaded dataset - shapefile name -, instead of the name of the
table imported into postgis.