[datafeeder] Allow overriding native CRS on getBounds/getSampleFeature
When an uploaded shapefile doesn't provide a .prj file, reprojecting to
the target CRS is impossible.

This API change allows the getBounds and getSampleFeature end points to
provide a "srsOverride" query parameter in addition to the "srs"
parameter, in order to define the source and target CRS for the
reprojection.

It also simplifies the semantics by removing the "srs_reproject"
parameter:

- If the "srs" parameter is provided, reprojection DOES HAPPEN.
- If there's no native CRS, it can be overridden with the "srsOverride"
  parameter. Otherwise, an error is thrown since it's not possible to
reproject from an unknown CRS.