[datafeeder] Fix (workaround) issue with GeoServer REST API
Change GeoServer datastore names from 'datafeeder' to
'<workspace>_datafeeder'.

Using 'datafeeder' as the GeoServer data store name for all workspaces
should be possible, but GeoServer has a bug where the requested
workspace/datastore name pair is ignored, and tries to use the first
datastore with the requested name ignoring the requested workspace,
when, for example, posting a FeatureType.

This resulted in unpredictable errors when publishing from georchestra
accounts in different organizations.