ldap - introducing unique object identifier
OpenLDAP provides a UUID as extra attribute, but this is vendor specific
and unstable in case of dumping the LDAP using ldapsearch (not slapcat).

Since we have a hand on the georchestra objectClasses already
introduced, the idea is to introduce a (optional) ldap attribute which
follows the UUID syntax.

Tests: tested in the docker image. For existing LDAP trees, the LDAP
schema for geOrchestra will have to be updated.