df-frontend - harmonizing build
Backport of #3772 for geOrchestra 22.0.x.

For now we are building the docker image for the datafeeder UI into a
separate repository, and the generic war / debian package into the main
repository.

this PR aims to harmonize the way we build, by building the docker image
at the same place.

Tests: `mvn clean package docker:build -Pdocker` into datafeeder-ui
generated a docker image which is quite the same as
georchestra/datafeeder-frontend:latest, having a quick look inside it.

If this PR is accepted, then the github action located into the
geonetwork-ui fork (branch georchestra-datafeeder) won't be necessary
anymore, and there won't be any need of setting a tag there either.

This should not change anything to the other builds (war,deb).